move wb_logger and logging_mixin
