+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `emote.callbacks` has been converted to a package. Future built-in
   callbacks will not be re-exported from `emote.callbacks`, and should
   instead be imported from their internal location.
+- `emote.callbacks.LoggingMixin` is now in the `emote.mixins.logging` module instead.
 
 ### Bugfixes
